--- java/io/jackbradshaw/otter/scene/integration/ParticleEffectIntegrator.kt
@@ -1,3 +1,0 @@
-package io.jackbradshaw.otter.scene.integration
-
-class ParticleEffectIntegrator
