+++ java/io/jackbradshaw/otter/scene/primitives/ParticleEffectIntegrator.kt
@@ -1,0 +1,3 @@
+package io.jackbradshaw.otter.scene.primitives
+
+class ParticleEffectIntegrator
