+++ java/io/jackbradshaw/otter/scene/primitives/PhysicsColliderIntegrator.kt
@@ -1,4 +1,4 @@
-package io.jackbradshaw.otter.scene.integration
+package io.jackbradshaw.otter.scene.primitives
 
 import com.jme3.bullet.collision.PhysicsCollisionObject
 import io.jackbradshaw.otter.engine.core.EngineCore
